requirement.txt
asgiref==3.5.2
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.1.2
django-crispy-forms==1.14.0
django-tinymce==3.5.0
et-xmlfile==1.1.0
MarkupPy==1.14
mysql-connector-python==8.0.31
mysqlclient==2.1.1
odfpy==1.4.1
openpyxl==3.0.10
Pillow==9.3.0
protobuf==3.20.1
PyYAML==6.0
sqlparse==0.4.3
tablib==3.2.1
tzdata==2022.4
xlrd==2.0.1
xlwt==1.3.0
